+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Yarn now accepts sponsorships! Please give a look at our [OpenCollective](https:
 
 - Reporting for Git errors has been improved.
 - The resolution step now has a progress indicator.
+- The experimental ESM loader warning emitted by Node.js is now suppressed.
 
 ## 3.1.1